refactor: enhance mission pre-condition support
Instead of only checking pre-conditions at the start of the gamemode,
call them every time the mission loop selects the mission.
This allows for changing the set of missions in real-time without
terminating the mission function, which would otherwise cause
a 1-2min delay between missions.

This could be enhanced further by allowing the pre-condition to
pass arguments to the mission function, such as the location to spawn in.